Doc: Adaptation to new configuration via boxes.yml
The documentation has been completely redesigned to document the changes
in the structure of the Vagrantfile as well as the additional support
of the Vagrant provider libvirt.

The description of the no longer required individual options for the
number of nodes for a particular distribution has been completely
removed.

The note about bug #9134 (https://github.com/hashicorp/vagrant/issues/9134)
in connection with Ubuntu 17.10 and Vagrant up to version 2.0.2 has been
removed, as the new Vagrant versions no longer have this bug and
Ubuntu 17.10 (Artful Aardvark) has reached "End of Life".