Add hint for problem when parallel installing
The provider libvirt performs the installation of the virtual machines
in parallel. Sometimes it happens that Ansible Provisioner is running
on the master node before all Ansible clients are up and running. Thus,
the Ansible Provisioner sometimes can not reach all clients via SSH and
Ansible will fail for the affected clients. In such a case, the Ansible
Provisioner on the master node must be rerun when all clients are up and
running.

  VAGRANT_DEFAULT_PROVIDER=libvirt vagrant provision master