Remove separate IP address discovery for libvirt
For the provider libvirt, a separate functionality for detecting the
IP address is not necessary with the Hostmanager plugin. The built-in
function correctly recognizes the IP address. Therefore, the extra
code has been removed.

For the provider VirtualBox another function for the discovery of
the IP address was built in, because 'hostname -I' for example on Alpine
Linux systems does not work. Instead, it will now use:
"ip addr show eth1 | grep -v ':' | egrep -o '([0-9] + \.) {3} [0-9]+'".