Commented out Alpine Linux
Currently there is a problem with the provider Virtualbox and Alpine
Linux. Name resolution does not work because the /etc/resolv.conf file
is empty after startup (see also
https://github.com/hashicorp/vagrant/issues/9588). If I change the
network configuration line in Vagrantfile from
  override.vm.network "private_network", type: "dhcp"
to
  override.vm.network "private_network", type: "dhcp", auto_config:
  false
as recommended in issue #9588 then the DNS server is correctly entered
into the file /etc/resolv.con via DHCP. But with 'auto_config: false'
the second network interface isn't coming up and the master node can not
connect to the Alpine Linux client.

With provider libvirt it works well.