Set interpreter discovery to 'auto'
auto:
Detects the target OS platform, distribution, and version, then consults
a table listing the correct Python interpreter and path for each
platform/distribution/version. If an entry is found, uses the discovered
interpreter. If no entry is found, or the listed Python is not present
on the target host, searches a list of common Python interpreter paths
and uses the first one found; also issues a warning that future
installation of another Python interpreter could alter the one chosen.

Conclusion:
This means that by default Python 3 is used instead of Python 2.